% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{WindowsMT.memory}
\alias{WindowsMT.memory}
\title{List of object sizes}
\usage{
WindowsMT.memory()
}
\description{
Returns the object sizes of all objects in memory.
}
\examples{
\dontrun{
WindowsMT.memory() 
}

}
